Add a flag for removal in install.sh
In ~install.sh (installation script), we originally removed the
configuration file (~~/ov.cfg), but now that only happens if the
-R flag is passed.

+++ install.sh
@@ -3,6 +3,10 @@
 bash build.sh
 sudo cp openvim /usr/local/bin/
 
+if [ ! $1 == "-R" ]; then
+    exit
+fi
+
 if [ -f "~/.ov.cfg" ]; then
     rm ~/.ov.cfg
 fi
